pyproject.toml
[build-system]
requires = ["setuptools>=64,<70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plasmangenuity"
version = "3.0.0"
description = "Universal wireless mouse battery monitor for Linux with HyperX configuration support"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = [
    "mouse", "battery", "wireless", "bluetooth", "upower",
    "hyperx", "logitech", "razer", "hid", "linux", "plasma", "kde",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: System :: Hardware",
    "Topic :: Utilities",
]

dependencies = [
    "hidapi>=0.14.0",
]

[project.optional-dependencies]
tray = [
    "PyQt5>=5.15",
    "pyudev>=0.24",
    "dbus-python>=1.3.2",
]

[project.scripts]
plasmangenuity = "plasmangenuity.cli:main"
plasmangenuity-tray = "plasmangenuity.tray:main"

[project.urls]
Homepage = "https://github.com/radoslavchobanov/plasmangenuity"
Issues = "https://github.com/radoslavchobanov/plasmangenuity/issues"

[tool.setuptools.packages.find]
include = ["plasmangenuity*"]
